meta: set msrv to 1.40

+++ iui/src/draw/strokeparams.rs
@@ -1,7 +1,6 @@
 use draw::DrawContext;
 use std::marker::PhantomData;
 use std::os::raw::c_double;
-use ui::UI;
 use ui_sys::uiDrawStrokeParams;
 
 pub use ui_sys::uiDrawLineCap as LineCap;
@@ -32,7 +31,7 @@ impl StrokeParams {
                 Thickness: self.thickness,
                 MiterLimit: self.miter_limit,
                 Dashes: self.dashes.as_ptr() as *mut c_double,
-                NumDashes: self.dashes.len(),
+                NumDashes: self.dashes.len() as u64,
                 DashPhase: self.dash_phase,
             },
             phantom: PhantomData,
